feat: Plateau is now a singleton, project structure re-organised
