+++ src/main/java/entities/Entity.java
@@ -1,0 +1,4 @@
+package entities;
+
+public class Entity {
+}
