+++ src/main/java/entities/environment/PlateauSize.java
@@ -1,4 +1,4 @@
-package datatypes;
+package entities.environment;
 
 public record PlateauSize(int x, int y) {
     public static final int DEFAULT_X = 5;
