+++ src/main/java/entities/vehicles/Rover.java
@@ -1,0 +1,4 @@
+package entities.vehicles;
+
+public class Rover {
+}
