+++ src/main/java/entities/vehicles/RoverControlUnit.java
@@ -1,4 +1,5 @@
-package vehicles;
+package entities.vehicles;
 
 public abstract class RoverControlUnit {
+
 }
