--- src/main/java/environment/Plateau.java
@@ -1,4 +1,0 @@
-package environment;
-
-public class Plateau {
-}
