+++ src/main/java/input/Verification/Verifier.java
@@ -1,4 +1,4 @@
-package Verification;
+package input.Verification;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
