+++ src/main/java/input/parsing/InstructionParser.java
@@ -1,4 +1,4 @@
-package parsing;
+package input.parsing;
 
 import datatypes.Instruction;
