+++ src/main/java/input/parsing/Parser.java
@@ -1,4 +1,4 @@
-package parsing;
+package input.parsing;
 
 public abstract class Parser <T> {
     public abstract T parse(String input);
