--- src/main/java/vehicles/Rover.java
@@ -1,4 +1,0 @@
-package vehicles;
-
-public class Rover {
-}
