+++ src/test/java/Verification/InstructionInputVerifierTest.java
@@ -1,6 +1,6 @@
 package Verification;
 
-import org.junit.jupiter.api.BeforeEach;
+import input.Verification.InstructionInputVerifier;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
