+++ src/test/java/Verification/PlateauSizeInputVerifierTest.java
@@ -1,5 +1,6 @@
 package Verification;
 
+import input.Verification.PlateauSizeInputVerifier;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
