+++ src/test/java/Verification/RoverPositionInputVerifierTest.java
@@ -1,5 +1,6 @@
 package Verification;
 
+import input.Verification.RoverPositionInputVerifier;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
